basic page layout for WeeklySnapshotPage
